PM Evaluations updates and costofsales

+++ application/views/admin_new/role.php
@@ -28,25 +28,22 @@
       <div class="card">
         <div class="card-header flex-wrap border-0 pt-6 pb-0">
           <div class="card-title">
-            <h3 class="card-label">Rules</h3>
+            <h3 class="card-label">Roles</h3>
           </div>
           <div class="card-toolbar">
-            <form class="form" id="role_form" action="<?php echo base_url() ?>admin/role" method="post"
-              enctype="multipart/form-data">
+            <form class="form" id="role_form" action="<?php echo base_url() ?>admin/role" method="post" enctype="multipart/form-data">
               <div class="card-body">
 
                 <div class="form-group row">
 
-                  <label class="col-lg-2 col-form-label text-lg-right">Rule</label>
+                  <label class="col-lg-2 col-form-label text-lg-right">Role</label>
                   <div class="col-lg-3">
                     <input type="text" class="form-control" name="name" value="<?= $name ?? '' ?>">
                   </div>
 
                   <div class="col-lg-7" style="text-align: center;">
-                    <button class="btn btn-success mr-2" name="search" type="submit" value="search"><i
-                        class="la la-search"></i>Search</button>
-                    <button class="btn btn-warning mr-2" name="submitReset" type="submit" value="submitReset"><i
-                        class="la la-trash"></i>Clear
+                    <button class="btn btn-success mr-2" name="search" type="submit" value="search"><i class="la la-search"></i>Search</button>
+                    <button class="btn btn-warning mr-2" name="submitReset" type="submit" value="submitReset"><i class="la la-trash"></i>Clear
                       Filter</button>
                   </div>
                 </div>
@@ -54,35 +51,31 @@
             </form>
             <!--begin::Button-->
             <?php if ($permission->add == 1) {
-              ?>
+            ?>
               <a href=" <?= base_url() ?>admin/addRole" class="btn btn-primary font-weight-bolder">
               <?php }
-            ?>
+              ?>
               <span class="svg-icon svg-icon-md">
-                <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="24px"
-                  height="24px" viewBox="0 0 24 24" version="1.1">
+                <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="24px" height="24px" viewBox="0 0 24 24" version="1.1">
                   <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
                     <rect x="0" y="0" width="24" height="24" />
                     <circle fill="#000000" cx="9" cy="15" r="6" />
-                    <path
-                      d="M8.8012943,7.00241953 C9.83837775,5.20768121 11.7781543,4 14,4 C17.3137085,4 20,6.6862915 20,10 C20,12.2218457 18.7923188,14.1616223 16.9975805,15.1987057 C16.9991904,15.1326658 17,15.0664274 17,15 C17,10.581722 13.418278,7 9,7 C8.93357256,7 8.86733422,7.00080962 8.8012943,7.00241953 Z"
-                      fill="#000000" opacity="0.3" />
+                    <path d="M8.8012943,7.00241953 C9.83837775,5.20768121 11.7781543,4 14,4 C17.3137085,4 20,6.6862915 20,10 C20,12.2218457 18.7923188,14.1616223 16.9975805,15.1987057 C16.9991904,15.1326658 17,15.0664274 17,15 C17,10.581722 13.418278,7 9,7 C8.93357256,7 8.86733422,7.00080962 8.8012943,7.00241953 Z" fill="#000000" opacity="0.3" />
                   </g>
                 </svg>
                 <!--end::Svg Icon-->
-              </span>Add New Rule</a>
-            <!--end::Button-->
+              </span>Add New Role</a>
+              <!--end::Button-->
           </div>
         </div>
         <div class="card-body">
           <!--begin: Datatable-->
 
-          <table class="table table-separate table-head-custom table-checkable table-hover table2excel"
-            id="kt_datatable2">
+          <table class="table table-separate table-head-custom table-checkable table-hover table2excel" id="kt_datatable2">
             <thead>
               <tr>
                 <th>ID</th>
-                <th>Rule</th>
+                <th>Role</th>
                 <th>Permission</th>
                 <th>Edit</th>
                 <th>Delete</th>
@@ -92,7 +85,7 @@
               <?php
               if ($role->num_rows() > 0) {
                 foreach ($role->result() as $row) {
-                  ?>
+              ?>
                   <tr class="">
                     <td>
                       <?php echo $row->id; ?>
@@ -101,7 +94,7 @@
                       <?php echo $row->name; ?>
                     </td>
                     <td><a href="<?php echo base_url() ?>admin/rolePermissions/<?=
-                         $row->id ?>" class="">
+                                                                                $row->id ?>" class="">
                         <i class="fa fa-cogs"></i>
                       </a></td>
 
@@ -110,29 +103,28 @@
                         <a href="<?php echo base_url() ?>admin/editRole/<?php echo $row->id; ?>" class="">
                           <i class="fa fa-pencil"></i> Edit
                         </a>
-                        <?php
+                      <?php
                       }
                       ?>
                     </td>
 
                     <td>
                       <?php if ($permission->delete == 1) { ?>
-                        <a href="<?php echo base_url() ?>admin/deleteRole/<?php echo $row->id ?>" title="delete" class=""
-                          onclick="return confirm('Are you sure you want to delete this user?');">
+                        <a href="<?php echo base_url() ?>admin/deleteRole/<?php echo $row->id ?>" title="delete" class="" onclick="return confirm('Are you sure you want to delete this user?');">
                           <i class="fa fa-times text-danger text"></i> Delete
                         </a>
-                        <?php
+                      <?php
                       }
                       ?>
                     </td>
                   </tr>
-                  <?php
+                <?php
                 }
               } else {
                 ?>
-              <tr>
-                <td colspan="7">There is no users to list</td>
-              </tr>
+                <tr>
+                  <td colspan="7">There is no users to list</td>
+                </tr>
               <?php
               }
               ?>
@@ -157,10 +149,10 @@
 
 
 <script>
-  $('#exportbtn').on('click', function () {
+  $('#exportbtn').on('click', function() {
     $(".table2excel").table2excel({
-      name: "Rules Sheet",
-      filename: "Rules",
+      name: "Roles Sheet",
+      filename: "Roles",
       fileext: ".xlsx",
       exclude_img: true,
       exclude_links: true,
@@ -169,5 +161,4 @@
     })
 
   })
-
 </script>
